requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
azure-core==1.31.0
azure-identity==1.19.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cohere==5.11.1
colorama==0.4.6
cryptography==43.0.3
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
eval_type_backport==0.2.0
fastavro==1.9.7
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.4
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.1
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpath-python==1.0.6
llama-cloud==0.1.4
llama-index==0.11.20
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.20
llama-index-embeddings-adapter==0.2.2
llama-index-embeddings-openai==0.2.5
llama-index-experimental==0.4.0
llama-index-finetuning==0.2.1
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-azure-openai==0.2.2
llama-index-llms-gemini==0.3.7
llama-index-llms-mistralai==0.2.7
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-postprocessor-cohere-rerank==0.2.1
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.11
MarkupSafe==3.0.2
marshmallow==3.23.0
mistralai==1.1.0
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.52.2
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
pillow==10.4.0
portalocker==2.10.1
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.2.0
pypdf==4.3.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.46.0
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wrapt==1.16.0
yarl==1.16.0
